gui-libs/xdg-desktop-portal-hyprland: add 9999 * add me as proxy mantainer

+++ gui-libs/xdg-desktop-portal-hyprland/xdg-desktop-portal-hyprland-1.2.2-r1.ebuild
@@ -8,12 +8,17 @@ inherit systemd cmake toolchain-funcs
 DESCRIPTION="xdg-desktop-portal backend for hyprland"
 HOMEPAGE="https://github.com/hyprwm/xdg-desktop-portal-hyprland"
 
-KEYWORDS="~amd64"
-PROTO_COMMIT="4d29e48433270a2af06b8bc711ca1fe5109746cd"
-SRC_URI="https://github.com/hyprwm/xdg-desktop-portal-hyprland/archive/refs/tags/v${PV}.tar.gz \
-	-> xdg-desktop-hyprland-${PV}.tar.gz
-https://github.com/hyprwm/hyprland-protocols/archive/${PROTO_COMMIT}.tar.gz \
-	-> proto-subproject-${PV}.tar.gz"
+if [[ ${PV} == 9999 ]]; then
+	EGIT_REPO_URI="https://github.com/hyprwm/xdg-desktop-portal-hyprland.git"
+	inherit git-r3
+else
+	PROTO_COMMIT="4d29e48433270a2af06b8bc711ca1fe5109746cd"
+	SRC_URI="https://github.com/hyprwm/xdg-desktop-portal-hyprland/archive/refs/tags/v${PV}.tar.gz \
+		-> xdg-desktop-hyprland-${PV}.tar.gz
+	https://github.com/hyprwm/hyprland-protocols/archive/${PROTO_COMMIT}.tar.gz \
+		-> proto-subproject-${PV}.tar.gz"
+	KEYWORDS="~amd64"
+fi
 
 LICENSE="MIT"
 SLOT="0"
@@ -67,9 +72,13 @@ pkg_setup() {
 }
 
 src_unpack() {
-	default
-	rmdir "${S}/subprojects/hyprland-protocols" || die
-	mv "hyprland-protocols-${PROTO_COMMIT}" "${S}/subprojects/hyprland-protocols" || die
+	if [[ ${PV} == 9999 ]]; then
+		git-r3_src_unpack
+	else
+		default
+		rmdir "${S}/subprojects/hyprland-protocols" || die
+		mv "hyprland-protocols-${PROTO_COMMIT}" "${S}/subprojects/hyprland-protocols" || die
+	fi
 }
 
 src_prepare() {
